+++ hivemq-edge/src/frontend/src/components/rjsf/ArrayFieldTemplate.tsx
@@ -4,7 +4,7 @@ import { RJSFSchema } from '@rjsf/utils/src/types.ts'
 import { Box, Grid, GridItem, HStack } from '@chakra-ui/react'
 
 import AddButton from '@/components/rjsf/__internals/AddButton.tsx'
-import BatchUploadButton from '@/components/rjsf/BatchSubscription/BatchUploadButton.tsx'
+import BatchUploadButton from '@/components/rjsf/BatchModeMappings/BatchUploadButton.tsx'
 import { AdapterContext } from '@/modules/ProtocolAdapters/types.ts'
 
 export const ArrayFieldTemplate: FC<ArrayFieldTemplateProps<unknown, RJSFSchema, AdapterContext>> = (props) => {
