+++ hivemq-edge/src/frontend/src/components/rjsf/BatchModeMappings/hooks/useBatchModeSteps.spec.ts
@@ -3,9 +3,9 @@ import { act, renderHook } from '@testing-library/react'
 
 import '@/config/i18n.config.ts'
 
-import { BatchModeStepType, BatchModeSteps, BatchModeStore } from '@/components/rjsf/BatchSubscription/types.ts'
-import { useBatchModeSteps } from '@/components/rjsf/BatchSubscription/hooks/useBatchModeSteps.ts'
-import { MOCK_ID_SCHEMA } from '@/components/rjsf/BatchSubscription/__test-utils__/store.mocks.ts'
+import { BatchModeStepType, BatchModeSteps, BatchModeStore } from '@/components/rjsf/BatchModeMappings/types.ts'
+import { useBatchModeSteps } from '@/components/rjsf/BatchModeMappings/hooks/useBatchModeSteps.ts'
+import { MOCK_ID_SCHEMA } from '@/components/rjsf/BatchModeMappings/__test-utils__/store.mocks.ts'
 
 const MOCK_STORE: BatchModeStore = {
   idSchema: MOCK_ID_SCHEMA,
