+++ hivemq-edge/src/frontend/src/components/rjsf/BatchModeMappings/types.ts
