+++ hivemq-edge/src/frontend/src/components/rjsf/BatchModeMappings/utils/config.utils.ts
