+++ hivemq-edge/src/frontend/src/components/rjsf/BatchModeMappings/utils/dropzone.utils.ts
