+++ hivemq-edge/src/frontend/src/components/rjsf/BatchModeMappings/utils/levenshtein.utils.ts
@@ -1,7 +1,7 @@
 // see https://github.com/gustf/js-levenshtein
 
-import { ColumnOption } from '@/components/rjsf/BatchSubscription/types.ts'
-import { AUTO_MATCH_DISTANCE } from '@/components/rjsf/BatchSubscription/utils/config.utils.ts'
+import { ColumnOption } from '@/components/rjsf/BatchModeMappings/types.ts'
+import { AUTO_MATCH_DISTANCE } from '@/components/rjsf/BatchModeMappings/utils/config.utils.ts'
 
 const _min = (d0: number, d1: number, d2: number, bx: number, ay: number): number => {
   return d0 < d1 || d2 < d1 ? (d0 > d2 ? d2 + 1 : d0 + 1) : bx === ay ? d1 : d1 + 1
